Refactoring seam hiding code to be extension methods.
