+++ Tests/MatterSlice.Tests/MatterSlice/PathOrderTests.cs
@@ -106,7 +106,7 @@ namespace MatterHackers.MatterSlice.Tests
 				// |      \
 				// |0______\1
 				List<IntPoint> testPoints = new List<IntPoint> { new IntPoint(0, 0), new IntPoint(100, 0), new IntPoint(70, 50), new IntPoint(100, 100), new IntPoint(0, 100) };
-				int bestPoint = PathOrderOptimizer.GetBestIndex(testPoints);
+				int bestPoint = testPoints.FindGreatestTurnIndex();
 				Assert.IsTrue(bestPoint == 2);
 			}
 
@@ -118,7 +118,7 @@ namespace MatterHackers.MatterSlice.Tests
 				// |       |
 				// |0______|1
 				List<IntPoint> testPoints = new List<IntPoint> { new IntPoint(0, 0), new IntPoint(100, 0), new IntPoint(100, 100), new IntPoint(0, 100) };
-				int bestPoint = PathOrderOptimizer.GetBestIndex(testPoints);
+				int bestPoint = testPoints.FindGreatestTurnIndex();
 				Assert.IsTrue(bestPoint == 3);
 			}
 
@@ -130,7 +130,7 @@ namespace MatterHackers.MatterSlice.Tests
 				// |       |
 				// |2______|3
 				List<IntPoint> testPoints = new List<IntPoint> { new IntPoint(100, 100), new IntPoint(0, 100), new IntPoint(0, 0), new IntPoint(100, 0) };
-				int bestPoint = PathOrderOptimizer.GetBestIndex(testPoints);
+				int bestPoint = testPoints.FindGreatestTurnIndex();
 				Assert.IsTrue(bestPoint == 1);
 			}
 
@@ -142,7 +142,7 @@ namespace MatterHackers.MatterSlice.Tests
 				// |       |
 				// |0______|3
 				List<IntPoint> testPoints = new List<IntPoint> { new IntPoint(0, 0), new IntPoint(0, 100), new IntPoint(100, 100), new IntPoint(100, 0) };
-				int bestPoint = PathOrderOptimizer.GetBestIndex(testPoints);
+				int bestPoint = testPoints.FindGreatestTurnIndex();
 				Assert.IsTrue(bestPoint == 1);
 			}
 
@@ -154,7 +154,7 @@ namespace MatterHackers.MatterSlice.Tests
 				// |       |
 				// |3______|2
 				List<IntPoint> testPoints = new List<IntPoint> { new IntPoint(0, 100), new IntPoint(100, 100), new IntPoint(100, 0), new IntPoint(0, 0) };
-				int bestPoint = PathOrderOptimizer.GetBestIndex(testPoints);
+				int bestPoint = testPoints.FindGreatestTurnIndex();
 				Assert.IsTrue(bestPoint == 0);
 			}
 
@@ -166,7 +166,7 @@ namespace MatterHackers.MatterSlice.Tests
 				// |      \
 				// |0______\1
 				List<IntPoint> testPoints = new List<IntPoint> { new IntPoint(0, 0), new IntPoint(1000, 0), new IntPoint(900, 500), new IntPoint(1000, 1000), new IntPoint(0, 1000) };
-				int bestPoint = PathOrderOptimizer.GetBestIndex(testPoints);
+				int bestPoint = testPoints.FindGreatestTurnIndex();
 				Assert.IsTrue(bestPoint == 2);
 			}
 
@@ -178,7 +178,7 @@ namespace MatterHackers.MatterSlice.Tests
 				// |      \
 				// |3______\4
 				List<IntPoint> testPoints = new List<IntPoint> { new IntPoint(90, 50), new IntPoint(100, 100), new IntPoint(0, 100), new IntPoint(0, 0), new IntPoint(100, 0) };
-				int bestPoint = PathOrderOptimizer.GetBestIndex(testPoints);
+				int bestPoint = testPoints.FindGreatestTurnIndex();
 				Assert.IsTrue(bestPoint == 0);
 			}
 
@@ -190,7 +190,7 @@ namespace MatterHackers.MatterSlice.Tests
 				//   /    \
 				//  /4_____\5
 				List<IntPoint> testPoints = new List<IntPoint> { new IntPoint(90, 50), new IntPoint(100, 100), new IntPoint(0, 100), new IntPoint(10, 50), new IntPoint(0, 0), new IntPoint(100, 0) };
-				int bestPoint = PathOrderOptimizer.GetBestIndex(testPoints);
+				int bestPoint = testPoints.FindGreatestTurnIndex();
 				Assert.IsTrue(bestPoint == 3);
 			}
 
@@ -202,7 +202,7 @@ namespace MatterHackers.MatterSlice.Tests
 				//   /    \
 				//  /4_____\5
 				List<IntPoint> testPoints = new List<IntPoint> { new IntPoint(950, 500), new IntPoint(1000, 1000), new IntPoint(0, 1000), new IntPoint(100, 500), new IntPoint(0, 0), new IntPoint(1000, 0) };
-				int bestPoint = PathOrderOptimizer.GetBestIndex(testPoints);
+				int bestPoint = testPoints.FindGreatestTurnIndex();
 				Assert.IsTrue(bestPoint == 3);
 			}
 
@@ -214,7 +214,7 @@ namespace MatterHackers.MatterSlice.Tests
 				//   /    \
 				//  /4_____\5
 				List<IntPoint> testPoints = new List<IntPoint> { new IntPoint(550, 500), new IntPoint(1000, 1000), new IntPoint(0, 1000), new IntPoint(100, 500), new IntPoint(0, 0), new IntPoint(1000, 0) };
-				int bestPoint = PathOrderOptimizer.GetBestIndex(testPoints);
+				int bestPoint = testPoints.FindGreatestTurnIndex();
 				Assert.IsTrue(bestPoint == 0);
 			}
 
@@ -226,7 +226,7 @@ namespace MatterHackers.MatterSlice.Tests
 				//   /    \
 				//  /1_____\2
 				List<IntPoint> testPoints = new List<IntPoint> { new IntPoint(10, 50), new IntPoint(0, 0), new IntPoint(100, 0), new IntPoint(90, 50), new IntPoint(100, 100), new IntPoint(0, 100), };
-				int bestPoint = PathOrderOptimizer.GetBestIndex(testPoints);
+				int bestPoint = testPoints.FindGreatestTurnIndex();
 				Assert.IsTrue(bestPoint == 0);
 			}
 
@@ -238,7 +238,7 @@ namespace MatterHackers.MatterSlice.Tests
 				// |      \
 				// |0______\4
 				List<IntPoint> testPoints = new List<IntPoint> { new IntPoint(0, 0), new IntPoint(0, 100), new IntPoint(100, 100), new IntPoint(90, 50), new IntPoint(100, 0) };
-				int bestPoint = PathOrderOptimizer.GetBestIndex(testPoints);
+				int bestPoint = testPoints.FindGreatestTurnIndex();
 				Assert.IsTrue(bestPoint == 1);
 			}
 
@@ -250,7 +250,7 @@ namespace MatterHackers.MatterSlice.Tests
 				// |      \
 				// |1______\0
 				List<IntPoint> testPoints = new List<IntPoint> { new IntPoint(100, 0), new IntPoint(0, 0), new IntPoint(0, 100), new IntPoint(100, 100), new IntPoint(90, 50) };
-				int bestPoint = PathOrderOptimizer.GetBestIndex(testPoints);
+				int bestPoint = testPoints.FindGreatestTurnIndex();
 				Assert.IsTrue(bestPoint == 2);
 			}
 
@@ -265,7 +265,7 @@ namespace MatterHackers.MatterSlice.Tests
 				{
 					new IntPoint(90, 50), new IntPoint(100, 0), new IntPoint(0, 0), new IntPoint(10, 50), new IntPoint(0, 100), new IntPoint(100, 100)
 				};
-				int bestPoint = PathOrderOptimizer.GetBestIndex(testPoints);
+				int bestPoint = testPoints.FindGreatestTurnIndex();
 				Assert.IsTrue(bestPoint == 4);
 			}
 
@@ -280,7 +280,7 @@ namespace MatterHackers.MatterSlice.Tests
 				{
 					new IntPoint(10, 50), new IntPoint(0, 100), new IntPoint(100, 100), new IntPoint(90, 50), new IntPoint(100, 0), new IntPoint(0, 0),
 				};
-				int bestPoint = PathOrderOptimizer.GetBestIndex(testPoints);
+				int bestPoint = testPoints.FindGreatestTurnIndex();
 				Assert.IsTrue(bestPoint == 1);
 			}
 		}
